edit: LoadAction class reads a file and draw.
the LoadAction class read the file then it reads the type of the
shape. After that it creates a pointer of the type of the shape
and a pointer of the type of CFigure. Then it assign the adress of the
shape to CFigure. Finally, using polymorphism, it calls Load function.
After that, it calls the add figure to add the shape and DrawAll
function to clear the draw area and draw the shapes.